Simplify: add simple API for users
